Zavrsen rad na permisijama i sredjivanje bagova.

+++ backend/src/main/java/com/project/Rentingaccommodation/controller/CityController.java
@@ -6,18 +6,26 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.project.Rentingaccommodation.model.City;
+import com.project.Rentingaccommodation.model.UserPrivileges;
+import com.project.Rentingaccommodation.model.UserRoles;
+import com.project.Rentingaccommodation.security.JwtUserPermissions;
 import com.project.Rentingaccommodation.service.CityService;
 
 @RestController
 @RequestMapping(value="/api/cities")
 public class CityController {
+	
 	@Autowired
 	private CityService service;
+	
+	@Autowired
+	private JwtUserPermissions jwtUserPermissions;
 	
 	@RequestMapping(value="", method=RequestMethod.GET)
 	public ResponseEntity<List<City>> getCities() {
@@ -44,7 +52,10 @@ public class CityController {
 	}
 	
 	@RequestMapping(value="", method=RequestMethod.POST)
-	public ResponseEntity<Object> addCity(@RequestBody City city) {
+	public ResponseEntity<Object> addCity(@RequestBody City city, @RequestHeader(value="Authorization") String token) {
+		if(!jwtUserPermissions.hasRoleAndPrivilege(token, UserRoles.ADMIN, UserPrivileges.WRITE_PRIVILEGE)) {
+			return new ResponseEntity<>("Not authorized", HttpStatus.UNAUTHORIZED);
+		}
 		if (city.getName() == null || city.getName() == "" ||
 			city.getZipcode() == null || city.getZipcode() == "" ||
 			city.getCountry() == null) {
@@ -60,7 +71,10 @@ public class CityController {
 	}
 	
 	@RequestMapping(value="/{id}", method=RequestMethod.PUT)
-	public ResponseEntity<Object> updateCity(@PathVariable Long id, @RequestBody City city) {
+	public ResponseEntity<Object> updateCity(@PathVariable Long id, @RequestBody City city, @RequestHeader(value="Authorization") String token) {
+		if(!jwtUserPermissions.hasRoleAndPrivilege(token, UserRoles.ADMIN, UserPrivileges.WRITE_PRIVILEGE)) {
+			return new ResponseEntity<>("Not authorized", HttpStatus.UNAUTHORIZED);
+		}
 		if (city.getName() == null || city.getName() == "" ||
 			city.getZipcode() == null || city.getZipcode() == "" ||
 			city.getCountry() == null) {
@@ -78,7 +92,10 @@ public class CityController {
 	}
 	
 	@RequestMapping(value="/{id}", method=RequestMethod.DELETE)
-	public ResponseEntity<Object> deleteCity(@PathVariable Long id) {
+	public ResponseEntity<Object> deleteCity(@PathVariable Long id, @RequestHeader(value="Authorization") String token) {
+		if(!jwtUserPermissions.hasRoleAndPrivilege(token, UserRoles.ADMIN, UserPrivileges.WRITE_PRIVILEGE)) {
+			return new ResponseEntity<>("Not authorized", HttpStatus.UNAUTHORIZED);
+		}
 		City city = service.delete(id);
 		if (city == null) {
 			return new ResponseEntity<>("City not found.", HttpStatus.NOT_FOUND);
