+++ backend/src/main/java/com/project/Rentingaccommodation/controller/CountryController.java
@@ -7,11 +7,15 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.project.Rentingaccommodation.model.Country;
+import com.project.Rentingaccommodation.model.UserPrivileges;
+import com.project.Rentingaccommodation.model.UserRoles;
+import com.project.Rentingaccommodation.security.JwtUserPermissions;
 import com.project.Rentingaccommodation.service.CountryService;
 
 @RestController
@@ -20,6 +24,9 @@ public class CountryController {
 	
 	@Autowired
 	private CountryService service;
+	
+	@Autowired
+	private JwtUserPermissions jwtUserPermissions;
 	
 	@RequestMapping(value="", method=RequestMethod.GET)
 	public ResponseEntity<List<Country>> getCountries() {
@@ -37,7 +44,10 @@ public class CountryController {
 	}
 	
 	@RequestMapping(value="", method=RequestMethod.POST)
-	public ResponseEntity<Object> addCountry(@RequestBody Country country) {
+	public ResponseEntity<Object> addCountry(@RequestBody Country country, @RequestHeader(value="Authorization") String token) {
+		if(!jwtUserPermissions.hasRoleAndPrivilege(token, UserRoles.ADMIN, UserPrivileges.WRITE_PRIVILEGE)) {
+			return new ResponseEntity<>("Not authorized", HttpStatus.UNAUTHORIZED);
+		}
 		if (country.getCode() == null || country.getCode() == "" ||
 			country.getName() == null || country.getName() == "") {
 			return new ResponseEntity<>("All fields are required.", HttpStatus.FORBIDDEN);
@@ -51,7 +61,10 @@ public class CountryController {
 	}
 	
 	@RequestMapping(value="/{id}", method=RequestMethod.PUT)
-	public ResponseEntity<Object> updateCountry(@PathVariable Long id, @RequestBody Country country) {
+	public ResponseEntity<Object> updateCountry(@PathVariable Long id, @RequestBody Country country, @RequestHeader(value="Authorization") String token) {
+		if(!jwtUserPermissions.hasRoleAndPrivilege(token, UserRoles.ADMIN, UserPrivileges.WRITE_PRIVILEGE)) {
+			return new ResponseEntity<>("Not authorized", HttpStatus.UNAUTHORIZED);
+		}
 		if (country.getCode() == null || country.getCode() == "" ||
 			country.getName() == null || country.getName() == "") {
 			return new ResponseEntity<>("All fields are required (code, name).", HttpStatus.FORBIDDEN);
@@ -67,7 +80,10 @@ public class CountryController {
 	}
 	
 	@RequestMapping(value="/{id}", method=RequestMethod.DELETE)
-	public ResponseEntity<Object> deleteCountry(@PathVariable Long id) {
+	public ResponseEntity<Object> deleteCountry(@PathVariable Long id, @RequestHeader(value="Authorization") String token) {
+		if(!jwtUserPermissions.hasRoleAndPrivilege(token, UserRoles.ADMIN, UserPrivileges.WRITE_PRIVILEGE)) {
+			return new ResponseEntity<>("Not authorized", HttpStatus.UNAUTHORIZED);
+		}
 		Country country = service.delete(id);
 		if (country == null) {
 			return new ResponseEntity<>("Country not found.", HttpStatus.NOT_FOUND);
